test: Added unittest for model_from_hf_hub
